\name{options.rts.l}
\alias{options.rts.l}

\title{
Provides the rts (returns to scale) option.
}

\description{
Provides the rts (returns to scale) option values.
}

\keyword{ DEA }
\keyword{ Multiplier DEA Model }
\keyword{ Cross Efficiency }
\keyword{ Single-Phase Approach }
\keyword{ Weight restrictions }
\keyword{ Assurance Region }
\keyword{ Malevolent }
\keyword{ Benevolent }
\keyword{ Two-Phase Approach }% __ONLY ONE__ keyword per line

