\name{dict.solveStatus}
\alias{dict.solveStatus}

\title{
Provides the solver status codes.
}

\description{
Provides the solver status codes and description.
}

\keyword{ DEA }
\keyword{ Multiplier DEA Model }
\keyword{ Cross Efficiency }
\keyword{ Single-Phase Approach }
\keyword{ Weight restrictions }
\keyword{ Assurance Region }
\keyword{ Malevolent }
\keyword{ Benevolent }
\keyword{ Two-Phase Approach }% __ONLY ONE__ keyword per line

