\name{options.phase.l}
\alias{options.phase.l}

\title{
Provides the second phase options.
}

\description{
Provides the second phase options available for Mal_Ben function.
}

\examples{
# List the phase option used as arguments.

options.phase.l
}

\keyword{ DEA }
\keyword{ Multiplier DEA Model }
\keyword{ Cross Efficiency }
\keyword{ Single-Phase Approach }
\keyword{ Weight restrictions }
\keyword{ Assurance Region }
\keyword{ Malevolent }
\keyword{ Benevolent }
\keyword{ Two-Phase Approach }% __ONLY ONE__ keyword per line



